refactor: linear algebra functions
